+++ EnhanceQoL/Settings/ChatFrame.lua
@@ -51,12 +51,14 @@ local function getChatIMSoundDropdownOptions()
 	return chatIMSoundOptions
 end
 
-local cChatFrame = addon.SettingsLayout.rootSOCIAL
+local cChatFrame = nil
 addon.SettingsLayout.chatframeCategory = cChatFrame
 
 local chatWindowExpandable = addon.functions.SettingsCreateExpandableSection(cChatFrame, {
 	name = L["ChatWindow"] or "Chat Window",
 	configPageKey = "ChatWindow",
+	modernCategory = "social",
+	modernOnly = true,
 	iconKey = "chatwindow",
 	expanded = false,
 	colorizeTitle = false,
@@ -256,6 +258,8 @@ addon.functions.SettingsCreateCheckboxes(cChatFrame, data)
 
 local chatIMExpandable = addon.functions.SettingsCreateExpandableSection(cChatFrame, {
 	name = L["InstantMessenger"] or "Instant Messenger",
+	modernCategory = "social",
+	modernOnly = true,
 	iconKey = "instantmessenger",
 	expanded = false,
 	colorizeTitle = false,
@@ -263,15 +267,6 @@ local chatIMExpandable = addon.functions.SettingsCreateExpandableSection(cChatFr
 })
 
 addon.functions.SettingsCreateText(cChatFrame, "|cff99e599" .. L["RightClickCloseTab"] .. "|r", { parentSection = chatIMExpandable })
-addon.functions.SettingsCreateCheckbox(cChatFrame, {
-	var = "hideQuickJoinToast",
-	text = HIDE .. " " .. COMMUNITIES_NOTIFICATION_SETTINGS_DIALOG_QUICK_JOIN_LABEL,
-	func = function(v)
-		addon.db["hideQuickJoinToast"] = v
-		addon.functions.toggleQuickJoinToastButton(addon.db["hideQuickJoinToast"])
-	end,
-	parentSection = chatIMExpandable,
-})
 
 data = {
 	{
@@ -514,6 +509,8 @@ addon.functions.SettingsCreateButton(cChatFrame, data)
 local chatHistoryExpandable = addon.functions.SettingsCreateExpandableSection(cChatFrame, {
 	name = L["CH_TITLE_HISTORY"] or "Chat History",
 	configPageKey = "ChatHistory",
+	modernCategory = "social",
+	modernOnly = true,
 	iconKey = "chathistory",
 	expanded = false,
 	colorizeTitle = false,
@@ -867,6 +864,8 @@ addon.functions.SettingsCreateCheckboxes(cChatFrame, data)
 local chatBubblesExpandable = addon.functions.SettingsCreateExpandableSection(cChatFrame, {
 	name = L["ChatBubbles"] or "Chat Bubbles",
 	configPageKey = "ChatBubbles",
+	modernCategory = "social",
+	modernOnly = true,
 	iconKey = "chatbubbles",
 	expanded = false,
 	colorizeTitle = false,